requirements.txt
woocommerce
fastapi
uvicorn
requests
openai
colorama
python-dotenv
pymongo
motor
